backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.1",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "transformers>=4.36.0",
    "torch>=2.0.0",
    "sentencepiece>=0.1.99",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "openai>=1.0.0",
]
